add feature-clean-code-stage1
