+++ index.html
@@ -9,8 +9,10 @@
 </head>
 
 <body>
-    <script src="./js/interfaceMatrix.js"></script>
-    <script src="./js/proporcionAurea.js"></script>
+    <svg id="opera" width="350px" height="585px">
+        <text x="13" y="70">¡Bienvenido!</text>
+    </svg>
+    <script type="module" src="./js/controlador.js"></script>
 </body>
 
 </html>
